refactor(kubernetes): Push multi-manifest description down
The KubernetesMultiManifestOperationDescription class is only
used by the delete operation; let's push it down and figure out
what to extract to a base class if/when we need another
implementation.
